+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased] - 2017-07-05
+## [0.1.2] - 2017-07-05
+### Added
+- Usage instructions.
+
 ### Changed
 - Declare strict types in every file.
 - Imported namespace constants into `UUID` class.
@@ -15,5 +18,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release to GitHub.
 
+[0.1.2]: https://github.com/brightnucleus/keys/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/brightnucleus/keys/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/brightnucleus/keys/compare/v0.0.0...v0.1.0